Re-generate pigeon code
