+++ lib/src/pigeons/map_interfaces.dart
@@ -2223,143 +2223,140 @@ class MapInterfaces_PigeonCodec extends StandardMessageCodec {
     } else if (value is Feature) {
       buffer.putUint8(158);
       writeValue(buffer, value.encode());
-    } else if (value is PerformanceStatisticsOptions) {
+    } else if (value is GlyphsRasterizationOptions) {
       buffer.putUint8(159);
       writeValue(buffer, value.encode());
-    } else if (value is GlyphsRasterizationOptions) {
+    } else if (value is TileCoverOptions) {
       buffer.putUint8(160);
       writeValue(buffer, value.encode());
-    } else if (value is TileCoverOptions) {
+    } else if (value is MbxEdgeInsets) {
       buffer.putUint8(161);
       writeValue(buffer, value.encode());
-    } else if (value is MbxEdgeInsets) {
+    } else if (value is CameraOptions) {
       buffer.putUint8(162);
       writeValue(buffer, value.encode());
-    } else if (value is CameraOptions) {
+    } else if (value is CameraState) {
       buffer.putUint8(163);
       writeValue(buffer, value.encode());
-    } else if (value is CameraState) {
+    } else if (value is CameraBoundsOptions) {
       buffer.putUint8(164);
       writeValue(buffer, value.encode());
-    } else if (value is CameraBoundsOptions) {
+    } else if (value is CameraBounds) {
       buffer.putUint8(165);
       writeValue(buffer, value.encode());
-    } else if (value is CameraBounds) {
+    } else if (value is MapAnimationOptions) {
       buffer.putUint8(166);
       writeValue(buffer, value.encode());
-    } else if (value is MapAnimationOptions) {
+    } else if (value is CoordinateBounds) {
       buffer.putUint8(167);
       writeValue(buffer, value.encode());
-    } else if (value is CoordinateBounds) {
+    } else if (value is MapDebugOptions) {
       buffer.putUint8(168);
       writeValue(buffer, value.encode());
-    } else if (value is MapDebugOptions) {
+    } else if (value is TileCacheBudgetInMegabytes) {
       buffer.putUint8(169);
       writeValue(buffer, value.encode());
-    } else if (value is TileCacheBudgetInMegabytes) {
+    } else if (value is TileCacheBudgetInTiles) {
       buffer.putUint8(170);
       writeValue(buffer, value.encode());
-    } else if (value is TileCacheBudgetInTiles) {
+    } else if (value is MapOptions) {
       buffer.putUint8(171);
       writeValue(buffer, value.encode());
-    } else if (value is MapOptions) {
+    } else if (value is ScreenCoordinate) {
       buffer.putUint8(172);
       writeValue(buffer, value.encode());
-    } else if (value is ScreenCoordinate) {
+    } else if (value is ScreenBox) {
       buffer.putUint8(173);
       writeValue(buffer, value.encode());
-    } else if (value is ScreenBox) {
+    } else if (value is CoordinateBoundsZoom) {
       buffer.putUint8(174);
       writeValue(buffer, value.encode());
-    } else if (value is CoordinateBoundsZoom) {
+    } else if (value is Size) {
       buffer.putUint8(175);
       writeValue(buffer, value.encode());
-    } else if (value is Size) {
+    } else if (value is RenderedQueryOptions) {
       buffer.putUint8(176);
       writeValue(buffer, value.encode());
-    } else if (value is RenderedQueryOptions) {
+    } else if (value is SourceQueryOptions) {
       buffer.putUint8(177);
       writeValue(buffer, value.encode());
-    } else if (value is SourceQueryOptions) {
+    } else if (value is FeatureExtensionValue) {
       buffer.putUint8(178);
       writeValue(buffer, value.encode());
-    } else if (value is FeatureExtensionValue) {
+    } else if (value is LayerPosition) {
       buffer.putUint8(179);
       writeValue(buffer, value.encode());
-    } else if (value is LayerPosition) {
+    } else if (value is QueriedRenderedFeature) {
       buffer.putUint8(180);
       writeValue(buffer, value.encode());
-    } else if (value is QueriedRenderedFeature) {
+    } else if (value is QueriedSourceFeature) {
       buffer.putUint8(181);
       writeValue(buffer, value.encode());
-    } else if (value is QueriedSourceFeature) {
+    } else if (value is QueriedFeature) {
       buffer.putUint8(182);
       writeValue(buffer, value.encode());
-    } else if (value is QueriedFeature) {
+    } else if (value is FeaturesetFeatureId) {
       buffer.putUint8(183);
       writeValue(buffer, value.encode());
-    } else if (value is FeaturesetFeatureId) {
+    } else if (value is FeatureState) {
       buffer.putUint8(184);
       writeValue(buffer, value.encode());
-    } else if (value is FeatureState) {
+    } else if (value is _Interaction) {
       buffer.putUint8(185);
       writeValue(buffer, value.encode());
-    } else if (value is _Interaction) {
+    } else if (value is _InteractionPigeon) {
       buffer.putUint8(186);
       writeValue(buffer, value.encode());
-    } else if (value is _InteractionPigeon) {
+    } else if (value is FeaturesetDescriptor) {
       buffer.putUint8(187);
       writeValue(buffer, value.encode());
-    } else if (value is FeaturesetDescriptor) {
+    } else if (value is FeaturesetFeature) {
       buffer.putUint8(188);
       writeValue(buffer, value.encode());
-    } else if (value is FeaturesetFeature) {
+    } else if (value is MapContentGestureContext) {
       buffer.putUint8(189);
       writeValue(buffer, value.encode());
-    } else if (value is MapContentGestureContext) {
+    } else if (value is _RenderedQueryGeometry) {
       buffer.putUint8(190);
       writeValue(buffer, value.encode());
-    } else if (value is _RenderedQueryGeometry) {
+    } else if (value is ProjectedMeters) {
       buffer.putUint8(191);
       writeValue(buffer, value.encode());
-    } else if (value is ProjectedMeters) {
+    } else if (value is MercatorCoordinate) {
       buffer.putUint8(192);
       writeValue(buffer, value.encode());
-    } else if (value is MercatorCoordinate) {
+    } else if (value is StyleObjectInfo) {
       buffer.putUint8(193);
       writeValue(buffer, value.encode());
-    } else if (value is StyleObjectInfo) {
+    } else if (value is StyleProjection) {
       buffer.putUint8(194);
       writeValue(buffer, value.encode());
-    } else if (value is StyleProjection) {
+    } else if (value is FlatLight) {
       buffer.putUint8(195);
       writeValue(buffer, value.encode());
-    } else if (value is FlatLight) {
+    } else if (value is DirectionalLight) {
       buffer.putUint8(196);
       writeValue(buffer, value.encode());
-    } else if (value is DirectionalLight) {
+    } else if (value is AmbientLight) {
       buffer.putUint8(197);
       writeValue(buffer, value.encode());
-    } else if (value is AmbientLight) {
+    } else if (value is MbxImage) {
       buffer.putUint8(198);
       writeValue(buffer, value.encode());
-    } else if (value is MbxImage) {
+    } else if (value is ImageStretches) {
       buffer.putUint8(199);
       writeValue(buffer, value.encode());
-    } else if (value is ImageStretches) {
+    } else if (value is ImageContent) {
       buffer.putUint8(200);
       writeValue(buffer, value.encode());
-    } else if (value is ImageContent) {
+    } else if (value is TransitionOptions) {
       buffer.putUint8(201);
       writeValue(buffer, value.encode());
-    } else if (value is TransitionOptions) {
+    } else if (value is CanonicalTileID) {
       buffer.putUint8(202);
       writeValue(buffer, value.encode());
-    } else if (value is CanonicalTileID) {
-      buffer.putUint8(203);
-      writeValue(buffer, value.encode());
     } else if (value is StylePropertyValue) {
-      buffer.putUint8(204);
+      buffer.putUint8(203);
       writeValue(buffer, value.encode());
     } else {
       super.writeValue(buffer, value);
@@ -2460,96 +2457,94 @@ class MapInterfaces_PigeonCodec extends StandardMessageCodec {
       case 158:
         return Feature.decode(readValue(buffer)!);
       case 159:
-        return PerformanceStatisticsOptions.decode(readValue(buffer)!);
-      case 160:
         return GlyphsRasterizationOptions.decode(readValue(buffer)!);
-      case 161:
+      case 160:
         return TileCoverOptions.decode(readValue(buffer)!);
-      case 162:
+      case 161:
         return MbxEdgeInsets.decode(readValue(buffer)!);
-      case 163:
+      case 162:
         return CameraOptions.decode(readValue(buffer)!);
-      case 164:
+      case 163:
         return CameraState.decode(readValue(buffer)!);
-      case 165:
+      case 164:
         return CameraBoundsOptions.decode(readValue(buffer)!);
-      case 166:
+      case 165:
         return CameraBounds.decode(readValue(buffer)!);
-      case 167:
+      case 166:
         return MapAnimationOptions.decode(readValue(buffer)!);
-      case 168:
+      case 167:
         return CoordinateBounds.decode(readValue(buffer)!);
-      case 169:
+      case 168:
         return MapDebugOptions.decode(readValue(buffer)!);
-      case 170:
+      case 169:
         return TileCacheBudgetInMegabytes.decode(readValue(buffer)!);
-      case 171:
+      case 170:
         return TileCacheBudgetInTiles.decode(readValue(buffer)!);
-      case 172:
+      case 171:
         return MapOptions.decode(readValue(buffer)!);
-      case 173:
+      case 172:
         return ScreenCoordinate.decode(readValue(buffer)!);
-      case 174:
+      case 173:
         return ScreenBox.decode(readValue(buffer)!);
-      case 175:
+      case 174:
         return CoordinateBoundsZoom.decode(readValue(buffer)!);
-      case 176:
+      case 175:
         return Size.decode(readValue(buffer)!);
-      case 177:
+      case 176:
         return RenderedQueryOptions.decode(readValue(buffer)!);
-      case 178:
+      case 177:
         return SourceQueryOptions.decode(readValue(buffer)!);
-      case 179:
+      case 178:
         return FeatureExtensionValue.decode(readValue(buffer)!);
-      case 180:
+      case 179:
         return LayerPosition.decode(readValue(buffer)!);
-      case 181:
+      case 180:
         return QueriedRenderedFeature.decode(readValue(buffer)!);
-      case 182:
+      case 181:
         return QueriedSourceFeature.decode(readValue(buffer)!);
-      case 183:
+      case 182:
         return QueriedFeature.decode(readValue(buffer)!);
-      case 184:
+      case 183:
         return FeaturesetFeatureId.decode(readValue(buffer)!);
-      case 185:
+      case 184:
         return FeatureState.decode(readValue(buffer)!);
-      case 186:
+      case 185:
         return _Interaction.decode(readValue(buffer)!);
-      case 187:
+      case 186:
         return _InteractionPigeon.decode(readValue(buffer)!);
-      case 188:
+      case 187:
         return FeaturesetDescriptor.decode(readValue(buffer)!);
-      case 189:
+      case 188:
         return FeaturesetFeature.decode(readValue(buffer)!);
-      case 190:
+      case 189:
         return MapContentGestureContext.decode(readValue(buffer)!);
-      case 191:
+      case 190:
         return _RenderedQueryGeometry.decode(readValue(buffer)!);
-      case 192:
+      case 191:
         return ProjectedMeters.decode(readValue(buffer)!);
-      case 193:
+      case 192:
         return MercatorCoordinate.decode(readValue(buffer)!);
-      case 194:
+      case 193:
         return StyleObjectInfo.decode(readValue(buffer)!);
-      case 195:
+      case 194:
         return StyleProjection.decode(readValue(buffer)!);
-      case 196:
+      case 195:
         return FlatLight.decode(readValue(buffer)!);
-      case 197:
+      case 196:
         return DirectionalLight.decode(readValue(buffer)!);
-      case 198:
+      case 197:
         return AmbientLight.decode(readValue(buffer)!);
-      case 199:
+      case 198:
         return MbxImage.decode(readValue(buffer)!);
-      case 200:
+      case 199:
         return ImageStretches.decode(readValue(buffer)!);
-      case 201:
+      case 200:
         return ImageContent.decode(readValue(buffer)!);
-      case 202:
+      case 201:
         return TransitionOptions.decode(readValue(buffer)!);
-      case 203:
+      case 202:
         return CanonicalTileID.decode(readValue(buffer)!);
-      case 204:
+      case 203:
         return StylePropertyValue.decode(readValue(buffer)!);
       default:
         return super.readValueOfType(type, buffer);
@@ -3482,55 +3477,6 @@ class _MapInterface {
       MapInterfaces_PigeonCodec();
 
   final String pigeonVar_messageChannelSuffix;
-
-  Future<void> startPerformanceStatisticsCollection(
-      PerformanceStatisticsOptions options) async {
-    final String pigeonVar_channelName =
-        'dev.flutter.pigeon.mapbox_maps_flutter._MapInterface.startPerformanceStatisticsCollection$pigeonVar_messageChannelSuffix';
-    final BasicMessageChannel<Object?> pigeonVar_channel =
-        BasicMessageChannel<Object?>(
-      pigeonVar_channelName,
-      pigeonChannelCodec,
-      binaryMessenger: pigeonVar_binaryMessenger,
-    );
-    final List<Object?>? pigeonVar_replyList =
-        await pigeonVar_channel.send(<Object?>[options]) as List<Object?>?;
-    if (pigeonVar_replyList == null) {
-      throw _createConnectionError(pigeonVar_channelName);
-    } else if (pigeonVar_replyList.length > 1) {
-      throw PlatformException(
-        code: pigeonVar_replyList[0]! as String,
-        message: pigeonVar_replyList[1] as String?,
-        details: pigeonVar_replyList[2],
-      );
-    } else {
-      return;
-    }
-  }
-
-  Future<void> stopPerformanceStatisticsCollection() async {
-    final String pigeonVar_channelName =
-        'dev.flutter.pigeon.mapbox_maps_flutter._MapInterface.stopPerformanceStatisticsCollection$pigeonVar_messageChannelSuffix';
-    final BasicMessageChannel<Object?> pigeonVar_channel =
-        BasicMessageChannel<Object?>(
-      pigeonVar_channelName,
-      pigeonChannelCodec,
-      binaryMessenger: pigeonVar_binaryMessenger,
-    );
-    final List<Object?>? pigeonVar_replyList =
-        await pigeonVar_channel.send(null) as List<Object?>?;
-    if (pigeonVar_replyList == null) {
-      throw _createConnectionError(pigeonVar_channelName);
-    } else if (pigeonVar_replyList.length > 1) {
-      throw PlatformException(
-        code: pigeonVar_replyList[0]! as String,
-        message: pigeonVar_replyList[1] as String?,
-        details: pigeonVar_replyList[2],
-      );
-    } else {
-      return;
-    }
-  }
 
   Future<void> loadStyleURI(String styleURI) async {
     final String pigeonVar_channelName =
